convert to mvvm architecture
