+++ app/src/main/java/com/example/filmsearch/App.kt
@@ -1,0 +1,27 @@
+package com.example.filmsearch
+
+import android.app.Application
+import com.example.filmsearch.data.MainRepository
+import com.example.filmsearch.domain.Interactor
+
+class App: Application() {
+    lateinit var repo: MainRepository
+    lateinit var interactor: Interactor
+
+    override fun onCreate() {
+        super.onCreate()
+        //Инициализируем экземпляр App, через который будем получать доступ к остальным переменным
+        instance = this
+        //Инициализируем репозиторий
+        repo = MainRepository()
+        //Инициализируем интерактор
+        interactor = Interactor(repo)
+    }
+
+    companion object {
+        //Здесь статически хранится ссылка на экземпляр App
+        lateinit var instance: App
+            //Приватный сеттер, чтобы нельзя было в эту переменную присвоить что-либо другое
+            private set
+    }
+}
