+++ app/src/main/java/com/example/filmsearch/data/MainRepository.kt
@@ -1,0 +1,63 @@
+package com.example.filmsearch.data
+
+import com.example.filmsearch.R
+import com.example.filmsearch.domain.Film
+
+class MainRepository {
+    val filmsDataBase = listOf(
+        Film(
+            "Игра в кальмара",
+            R.drawable.play,
+            "Сотни игроков, испытывающих нехватку денег, принимают странное приглашение поучаствовать в детских играх. Внутри их ждет заманчивый приз со смертельно высокими ставками: игра на выживание, в которой на кону стоит колоссальный выигрыш в размере 45,6 миллиарда вон.",
+            6.5f
+        ),
+        Film(
+            "Первозданная америка",
+            R.drawable.america,
+            "Вас ждет увлекательный рассказ о смелом и авантюрном исследовании зарождения американского Запада, жестоких столкновениях культов, религий, мужчин и женщин,борющихся за контроль над новым миром.",
+            4.8f
+        ),
+        Film(
+            "Сегун",
+            R.drawable.segun,
+            "Когда в соседней японской рыбацкой деревушке находят таинственный европейский корабль, правитель Есии Торанага обнаруживает секреты, которые могут склонить чашу весов власти и уничтожить его врагов.",
+            7.9f
+        ),
+        Film(
+            "Субстанция",
+            R.drawable.substance,
+            "Увядающая знаменитость принимает наркотик с черного рынка: вещество, размножающее клетки, которое временно создает более молодую и совершенную версию самой себя.",
+            7.1f
+        ),
+        Film(
+            "Дюна",
+            R.drawable.duna,
+            "Пол Атрейдес прибывает на Арракис после того, как его отец принимает руководство опасной планетой. Однако после предательства наступает хаос, когда силы сталкиваются за контроль над меланжем, драгоценным ресурсом.",
+            6.9f
+        ),
+        Film(
+            "Паразиты",
+            R.drawable.cover,
+            "Жадность и классовая дискриминация угрожают недавно сложившимся симбиотическим отношениям между богатой семьей Пак и обездоленным кланом Ким.",
+            7.3f
+        ),
+        Film(
+            "Джокер",
+            R.drawable.jocer,
+            "Артур Флек, тусовочный клоун и неудавшийся стендап-комик, ведет бедную жизнь со своей больной матерью. Однако, когда общество избегает его и клеймит как урода, он решает принять жизнь хаоса в Готэм-Сити.",
+            7.6f
+        ),
+        Film(
+            "Побег из Шоушенка",
+            R.drawable.escape,
+            "Банкир, осужденный за убийство, на протяжении четверти века завязывает дружбу с закоренелым осужденным, сохраняя при этом свою невиновность и пытаясь сохранять надежду с помощью простого сострадания.",
+            8.0f
+        ),
+        Film(
+            "Властелин колец: Братство кольца",
+            R.drawable.lord,
+            "Кроткий Хоббит из Удела и восемь его спутников отправляются в путешествие, чтобы уничтожить могущественное Кольцо Единого и спасти Средиземье от Темного Лорда Саурона.",
+            8.9f
+        )
+    )
+}
