+++ app/src/main/java/com/example/filmsearch/domain/Interactor.kt
@@ -1,0 +1,7 @@
+package com.example.filmsearch.domain
+
+import com.example.filmsearch.data.MainRepository
+
+class Interactor(val repo: MainRepository) {
+    fun getFilmsDB(): List<Film> = repo.filmsDataBase
+}
