+++ app/src/main/java/com/example/filmsearch/viewmodel/HomeFragmentViewModel.kt
@@ -1,0 +1,18 @@
+package com.example.filmsearch.viewmodel
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import com.example.filmsearch.App
+import com.example.filmsearch.domain.Film
+import com.example.filmsearch.domain.Interactor
+
+class HomeFragmentViewModel: ViewModel() {
+    val filmsListLiveData = MutableLiveData<List<Film>>()
+    private var interactor: Interactor = App.instance.interactor
+    init {
+        //TODO init interactor
+
+        val films = interactor.getFilmsDB()
+        filmsListLiveData.postValue(films)
+    }
+}
