+++ app/src/main/res/layout/fragment_favorites.xml
@@ -7,7 +7,7 @@
     android:id="@+id/fragment_favorites"
     tools:visibility="visible"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".FavoritesFragment">
+    tools:context=".view.fragments.FavoritesFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/favorites_recycler"
