+++ app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/fragment_placeholder"
-    tools:context=".HomeFragment">
+    tools:context=".view.fragments.HomeFragment">
     <androidx.appcompat.widget.SearchView
         android:id="@+id/search_view"
         android:background="@drawable/search"
