+++ app/src/main/res/layout/fragment_selections.xml
@@ -4,7 +4,7 @@
     android:id="@+id/fragment_selections"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SelectionsFragment"
+    tools:context=".view.fragments.SelectionsFragment"
     tools:visibility="visible"
     android:background="@drawable/shape2">
